requirements.txt
torch
torchvision
coremltools
psutil
numpy
pillow